requirements.txt
openai
pdfminer.six
langchain-text-splitters
boto3
opensearch-py
requests-aws4auth
python-dotenv
pydantic
asyncio-extras
air